Credential encryption is platform, not AI — the protector moves and stops riding AddAI()
`IProviderKeyProtector` / `IMasterKeyProvider` / `ProviderKeyProtector` / `ConfigMasterKeyProvider`
move from `MeshWeaver.AI` to `MeshWeaver.Mesh.Contract`'s `MeshWeaver.Mesh.Security`. Step A2 of
#2276.

They were never AI. Four unrelated subsystems already encrypt with them: a model provider's
`ApiKey`, a GitHub PAT (`GitHubCredential`), the Entra EA credential (`EaGraphAuth`), and the plugin
catalog's sync-token signing key (`SyncTokenSigningKeyService`) — the last two reaching the type
through `GitSync`'s transitive reference to the AI engine.

**The registration had to move with them, and that is the load-bearing half.** It sat in
`AddModelProviderType()`, i.e. behind `AddAI()`. Every consumer resolves the protector OPTIONALLY —
`GetService`, and on null stores the secret verbatim — so the day the engine leaves for a module, a
deployment without it would have written all four credentials in PLAINTEXT and said nothing: no
exception, no warning, and the values still "work" because `Unprotect` passes untagged input
through. `AddGraph()` now registers it, so it exists on any mesh built from the platform.

`ProviderKeyProtectorRegistrationTest` is the ratchet, and it was verified to FAIL with the two
lines removed (`Expected value not to be <null>`) — which also proves nothing else in the tree
registers them.

The config key stays `Ai:KeyProtection:MasterKey` deliberately: it is live in Key Vault and in the
CSI SecretProviderClass mapping of every deployment, and DeploymentOptions.md records that a
DIFFERENT master key makes every stored `enc:` value undecryptable. A rename here would be a silent
deletion of a running secret, not a tidy-up.

Verified with the full solution build under CI's flags (`-c Release -p:CIRun=true -warnaserror`),
which is what surfaced the three doc `cref`s and the fully-qualified `MeshWeaver.AI.IProviderKeyProtector`
in the chat view. 23 provider-key tests plus the new ratchet pass.
