+++ memex/Memex.Portal.Shared/Models/ModelProviderService.cs
@@ -1,3 +1,4 @@
+using MeshWeaver.Mesh.Security;
 using System.Collections.Immutable;
 using System.Reactive;
 using System.Reactive.Linq;
