+++ src/MeshWeaver.AI/ModelProviderLayoutAreas.cs
@@ -1,3 +1,4 @@
+using MeshWeaver.Mesh.Security;
 using System.Collections.Immutable;
 using System.Linq;
 using System.Reactive.Linq;
