+++ src/MeshWeaver.AI/ModelStaticRepoSource.cs
@@ -1,3 +1,4 @@
+using MeshWeaver.Mesh.Security;
 using MeshWeaver.Markdown;
 using MeshWeaver.Mesh;
