+++ test/MeshWeaver.AI.Test/ProviderCredentialSeedDecisionTest.cs
@@ -1,5 +1,6 @@
 #pragma warning disable CS1591
 
+using MeshWeaver.Mesh.Security;
 using Xunit;
 
 namespace MeshWeaver.AI.Test;
