+++ test/MeshWeaver.AI.Test/ProviderKeyProtectorTest.cs
@@ -1,5 +1,6 @@
 #pragma warning disable CS1591
 
+using MeshWeaver.Mesh.Security;
 using System.Security.Cryptography;
 using MeshWeaver.AI;
 using Xunit;
