+++ test/MeshWeaver.GitSync.Test/GitHubSyncTestBase.cs
@@ -1,3 +1,4 @@
+using MeshWeaver.Mesh.Security;
 using System.Reactive.Linq;
 using System.Reactive.Threading.Tasks;
 using MeshWeaver.AI;
